Move hash_dict into data_util to use-use the NumpyEncoder that handle floats, fixes #1889
