+++ ludwig/features/feature_utils.py
@@ -20,7 +20,7 @@ import numpy as np
 import torch
 
 from ludwig.constants import NAME, PREPROCESSING, SEQUENCE, TEXT, TIMESERIES
-from ludwig.utils.misc_utils import hash_dict
+from ludwig.utils.data_utils import hash_dict
 from ludwig.utils.strings_utils import tokenizer_registry, UNKNOWN_SYMBOL
 
 SEQUENCE_TYPES = {SEQUENCE, TEXT, TIMESERIES}
